% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlexGebOR.R
\name{OR_95CI}
\alias{OR_95CI}
\title{Calculate Odds Ratio with Confidence Intervals}
\usage{
OR_95CI(coef, se, siglevel, roundto)
}
\arguments{
\item{coef}{Coefficient Estimate from model summary output}

\item{se}{Standard Error from model summary output}

\item{siglevel}{Desired alpha/significance level}

\item{roundto}{Preferred number of decimals for rounding}
}
\value{
OR with CI
}
\description{
Calculate Odds Ratio with Confidence Intervals
}
\examples{
OR_95CI(2,1,0.95,3)
"7.389 (6.940, 7.867)"

}
